move evm-testing to its own lib
