+++ libs/evm-testing/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "@hicommonwealth/evm-testing",
+  "private": "true",
+  "type": "module",
+  "version": "0.1.0",
+  "description": "A private blockchain environment with a private chain and transaction/interface abstraction API",
+  "files": [
+    "build"
+  ],
+  "main": "build/index.js",
+  "scripts": {
+    "build": "tsc -b ./tsconfig.build.json",
+    "clean": "rm -rf build && find . -type f -name '*.tsbuildinfo' -exec rm {} +",
+    "check-types": "tsc --noEmit",
+    "test": "echo No tests to run",
+    "start": "tsx src/server.ts",
+    "ganache": "node node_modules/ganache/dist/node/cli.js --fork --miner.blockTime 12 --wallet.unlockedAccounts 0xF977814e90dA44bFA03b6295A0616a897441aceC --wallet.unlockedAccounts 0xfA9b5f7fDc8AB34AAf3099889475d47febF830D7 --wallet.unlockedAccounts 0x6d903f6003cca6255D85CcA4D3B5E5146dC33925 --wallet.unlockedAccounts 0xBE0eB53F46cd790Cd13851d5EFf43D12404d33E8"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "axios": "^1.3.4",
+    "cors": "^2.8.5",
+    "dotenv": "^16.0.3",
+    "express": "^4.18.2",
+    "morgan": "^1.10.0",
+    "tsx": "^4.7.2",
+    "web3": "^4.7.0"
+  },
+  "optionalDependencies": {
+    "ganache": "^7.9.2"
+  }
+}
