+++ libs/evm-testing/src/utils/getProvider.ts
@@ -1,5 +1,5 @@
 import Web3 from 'web3';
-import { PROVIDER_URL } from '../../config';
+import { PROVIDER_URL } from '../config';
 
 const getProvider = () => {
   return new Web3(new Web3.providers.HttpProvider(String(PROVIDER_URL)));
