+++ libs/evm-testing/tsconfig.build.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "build"
+  },
+  "include": ["src", "src/sdk"],
+  "references": [{ "path": "../logging/tsconfig.build.json" }]
+}
