--- packages/commonwealth/test/util/evm-chain-testing/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "chain-testing",
-  "version": "1.0.0",
-  "type": "module",
-  "description": "A private blockchain environment with a private chain and transaction/interface abstraction API",
-  "main": "index.ts",
-  "scripts": {
-    "start": "tsx src/index.ts",
-    "ganache": "node node_modules/ganache/dist/node/cli.js --fork --miner.blockTime 12 --wallet.unlockedAccounts 0xF977814e90dA44bFA03b6295A0616a897441aceC --wallet.unlockedAccounts 0xfA9b5f7fDc8AB34AAf3099889475d47febF830D7 --wallet.unlockedAccounts 0x6d903f6003cca6255D85CcA4D3B5E5146dC33925 --wallet.unlockedAccounts 0xBE0eB53F46cd790Cd13851d5EFf43D12404d33E8",
-    "build": "tsc",
-    "test-sdk": "tsx test_sdk.ts",
-    "check-types": "tsc --noEmit"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "axios": "^1.3.4",
-    "cors": "^2.8.5",
-    "dotenv": "^16.0.3",
-    "express": "^4.18.2",
-    "morgan": "^1.10.0",
-    "tsx": "^4.7.2",
-    "web3": "^4.7.0"
-  },
-  "optionalDependencies": {
-    "ganache": "^7.9.2"
-  }
-}
