+++ tsconfig.json
@@ -4,6 +4,7 @@
     { "path": "libs/adapters/tsconfig.build.json" },
     { "path": "libs/chains/tsconfig.build.json" },
     { "path": "libs/core/tsconfig.build.json" },
+    { "path": "libs/evm-testing/tsconfig.build.json" },
     { "path": "libs/logging/tsconfig.build.json" },
     { "path": "libs/model/tsconfig.build.json" },
     { "path": "libs/shared/tsconfig.build.json" },
